COTURN: In the end, the problem was not the start script
It was the fact that the SQLite database was hard to overwrite.
See, the database is ALREADY installed during the Docker setup,
and if you try to place an overlay over it, things become weird,
since you don't know if `turnserver` is using the overlayed DB
or the previous one.

The solution :

* Provide a sample SQLite database in a volume
* Provide a turnserver.conf referencing this database

Problem solved...

It would have been better if I had thought about that before
writing a long-ass script to autogenerate turnserver.conf on the
fly... I gave up upon that, because I like to mount configuration
files as ReadOnly for security purposes.
I still kept the scripts and they're working, just in case I
find a usecase for them.

Signed-off-by: Miouyouyou (Myy) <myy@miouyouyou.fr>