Ok, pseudo automation should be almost done.
This is far from being completely automated... Buuut, this should be
able to create a working installation in a few scripts invocations.

You'll have to write the nice integration agent procedure, or the
"big script that invoke the small scripts" yourself though.

One of the reason is that, even though this is an Open Source project
with ZERO guarantee or warranties, some people might still throw a
tantrum if one those scripts behaved badly... instead of, you known,
identify where things went wrong, provide a fix and send a pull
request !

Signed-off-by: Miouyouyou (Myy) <myy@miouyouyou.fr>