This should do the trick.
The manual part of the README.md is fairly untested.

The automatic part is tested and working successfully.

I'm done with this shit for now. Matrix is an interesting project
but... let's be honest, if you're building a complex chat server
without a propre administration tool with a nice UI, your project
isn't ready.
I mean, just adding a user requires to invoke a random script with
THE RIGHT PARAMETERS. Don't you dare mess up with the parameters or
the thing will just NOT work.
I don't know how to delete a user and I don't know where to find
a good documentation on this subject so I'm giving up for the moment.

Signed-off-by: Miouyouyou (Myy) <myy@miouyouyou.fr>